First cut, using webpack to build react-tools.js
