+++ .gitignore
@@ -6,3 +6,4 @@ inst/doc
 node_modules
 reactR.Rcheck
 reactR_*.tar.gz
+*.swp
