+++ inst/www/react-tools/react-tools.js
@@ -1,120 +1,2 @@
-window.reactR = (function () {
-    /**
-     * Recursively transforms tag, a JSON representation of an instance of a
-     * React component and its children, into a React element suitable for
-     * passing to ReactDOM.render.
-     * @param {Object} components
-     * @param {Object} tag
-     */
-    function hydrate(components, tag) {
-        if (typeof tag === 'string') return tag;
-        if (tag.name[0] === tag.name[0].toUpperCase()
-            && !components.hasOwnProperty(tag.name)) {
-            throw new Error("Unknown component: " + tag.name);
-        }
-        var elem = components.hasOwnProperty(tag.name) ? components[tag.name] : tag.name,
-            args = [elem, tag.attribs];
-        for (var i = 0; i < tag.children.length; i++) {
-            args.push(hydrate(components, tag.children[i]));
-        }
-        return React.createElement.apply(React, args);
-    }
-
-    var defaultOptions = {
-        // The name of the property on the root tag to use for the width, if
-        // it's updated.
-        widthProperty: "width",
-        // The name of the property on the root tag to use for the height, if
-        // it's updated.
-        heightProperty: "height",
-        // Whether or not to append the string 'px' to the width and height
-        // properties when they change.
-        appendPx: false,
-        // Whether or not to dynamically update the width and height properties
-        // of the last known tag when the computed width and height change in
-        // the browser.
-        renderOnResize: false
-    };
-
-    function mergeOptions(options) {
-        var merged = {};
-        for (var k in defaultOptions) {
-            merged[k] = defaultOptions[k];
-        }
-        for (var k in options) {
-            if (!defaultOptions.hasOwnProperty(k)) {
-                throw new Error("Unrecognized option: " + k);
-            }
-            merged[k] = options[k];
-        }
-        return merged;
-    }
-
-    function formatDimension(dim, options) {
-        if (options.appendPx) {
-            return dim + 'px';
-        } else {
-            return dim;
-        }
-    }
-
-    function isTag(value) {
-        return (typeof value === 'object')
-            && value.hasOwnProperty('name')
-            && value.hasOwnProperty('attribs')
-            && value.hasOwnProperty('children');
-    }
-
-    /**
-     * Creates an HTMLWidget that is updated by rendering a React component.
-     * React component constructors are made available by specifying them by
-     * name in the components object.
-     * @param {string} name
-     * @param {string} type
-     * @param {Object} components
-     * @param {Object} options
-     */
-    function reactWidget(name, type, components, options) {
-        var actualOptions = mergeOptions(options);
-        HTMLWidgets.widget({
-            name: name,
-            type: type,
-            factory: function (el, width, height) {
-                var lastValue,
-                    renderValue = (function (value) {
-                        if (actualOptions.renderOnResize) {
-                            // value.tag might be a primitive string, in which
-                            // case there is no attribs property.
-                            if (typeof value.tag === 'object') {
-                                value.tag.attribs[actualOptions["widthProperty"]] = formatDimension(width);
-                                value.tag.attribs[actualOptions["heightProperty"]] = formatDimension(height);
-                            }
-                            lastValue = value;
-                        }
-                        ReactDOM.render(hydrate(components, value.tag), el);
-                    });
-                return {
-                    renderValue: renderValue,
-                    resize: function (newWidth, newHeight) {
-                        if (actualOptions.renderOnResize) {
-                            width = newWidth;
-                            height = newHeight;
-                            renderValue(lastValue);
-                        }
-                    }
-                };
-            }
-        })
-    }
-
-    return {
-        reactWidget: reactWidget,
-        hydrate: hydrate,
-        __internal: {
-            defaultOptions: defaultOptions,
-            mergeOptions: mergeOptions,
-            formatDimension: formatDimension,
-            isTag: isTag
-        }
-    };
-})()
+!function(e){var t={};function r(n){if(t[n])return t[n].exports;var o=t[n]={i:n,l:!1,exports:{}};return e[n].call(o.exports,o,o.exports,r),o.l=!0,o.exports}r.m=e,r.c=t,r.d=function(e,t,n){r.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:n})},r.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},r.t=function(e,t){if(1&t&&(e=r(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var n=Object.create(null);if(r.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var o in e)r.d(n,o,function(t){return e[t]}.bind(null,o));return n},r.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return r.d(t,"a",t),t},r.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},r.p="",r(r.s=0)}([function(e,t){function r(e){return(r="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}window.reactR=function(){function e(t,r){if("string"==typeof r)return r;if(r.name[0]===r.name[0].toUpperCase()&&!t.hasOwnProperty(r.name))throw new Error("Unknown component: "+r.name);for(var n=[t.hasOwnProperty(r.name)?t[r.name]:r.name,r.attribs],o=0;o<r.children.length;o++)n.push(e(t,r.children[o]));return React.createElement.apply(React,n)}var t={widthProperty:"width",heightProperty:"height",appendPx:!1,renderOnResize:!1};function n(e){var r={};for(var n in t)r[n]=t[n];for(var n in e){if(!t.hasOwnProperty(n))throw new Error("Unrecognized option: "+n);r[n]=e[n]}return r}function o(e,t){return t.appendPx?e+"px":e}return{reactWidget:function(t,i,a,u){var f=n(u);HTMLWidgets.widget({name:t,type:i,factory:function(t,n,i){var u,c=function(c){f.renderOnResize&&("object"===r(c.tag)&&(c.tag.attribs[f.widthProperty]=o(n),c.tag.attribs[f.heightProperty]=o(i)),u=c),ReactDOM.render(e(a,c.tag),t)};return{renderValue:c,resize:function(e,t){f.renderOnResize&&(n=e,i=t,c(u))}}}})},hydrate:e,__internal:{defaultOptions:t,mergeOptions:n,formatDimension:o,isTag:function(e){return"object"===r(e)&&e.hasOwnProperty("name")&&e.hasOwnProperty("attribs")&&e.hasOwnProperty("children")}}}}()}]);
+//# sourceMappingURL=react-tools.js.map
