+++ package.json
@@ -5,15 +5,16 @@
         "karma-chai": "^0.1.0",
         "karma-chrome-launcher": "^2.2.0",
         "karma-mocha": "^1.3.0",
-        "karma-webpack": "^3.0.5",
+        "karma-webpack": "^4.0.0-rc.6",
         "karma-source-map-support": "^1.3.0",
         "mocha": "^5.2.0",
         "react": "^16.7.0",
         "react-dom": "^16.7.0",
-        "webpack": "^3.0.0",
-        "@babel/core": "^7.2.0",
-        "babel-loader": "^8.0.4",
-        "@babel/preset-env": "^7.2.0",
+        "webpack": "^4.27.1",
+        "webpack-cli": "^3.1.2",
+        "@babel/core": "^7.2.2",
+        "babel-loader": "^8.0.5",
+        "@babel/preset-env": "^7.3.1",
         "@babel/preset-react": "^7.0.0",
         "@rgrove/parse-xml": "^1.1.1",
         "react-html-parser": "^2.0.2"
